Enhance payment handling and PcBuild features
- Added a fallback IP address ("127.0.0.1") in `PaymentController` to ensure a default value is used when `RemoteIpAddress` is null.
- Changed `GetPcBuildById` in `PcBuildController` to allow anonymous access by replacing `[Authorize]` with `[AllowAnonymous]`.
- Introduced a new `stockQuantity` property in `PcBuildItemResponse` to represent product stock levels.
- Updated `PcBuildItemMapper` to map the new `stockQuantity` property with a default value of `0` if null.
- Modified `PaymentService` to use `decimal` instead of `int` for `vnp_Amount` to improve precision in monetary calculations.

+++ Service/ImplementationsService/PaymentService.cs
@@ -67,7 +67,7 @@ namespace ShopPC.Service.ImplementationsService
                 { "vnp_Version", "2.1.0" },
                 { "vnp_Command", "pay" },
                 { "vnp_TmnCode", _vnpayConfig.TmnCode },
-                { "vnp_Amount", ((int)(amount * 100)).ToString() },
+                { "vnp_Amount", ((decimal)(amount * 100)).ToString() },
                 { "vnp_CreateDate", DateTime.Now.ToString("yyyyMMddHHmmss") },
                 { "vnp_CurrCode", "VND" },
                 { "vnp_IpAddr", ipAddress },
